feat: ModeStream (Redis-like append-only log)
Add ModeStream: XAdd/XRange/XRevRange/XLen/XDel/XTrim with owner-minted
millis-seq ids and a single opaque []byte payload. Snapshot fan-out
(FlagStream + S/A/D/T); non-owner XAdd uses peer StreamAdd.

Entry.Flags is uint64 so FlagStream can be 1<<32. XRevRange uses the
same start/end window as XRange (newest first). Hard cap 4096 returns
stream full. Client rejects int32 overflow on XTrim/count.

+++ examples/lab/http.go
@@ -286,7 +286,7 @@ func (l *Lab) snapshot(ks, key string) map[string]any {
 	}
 	out := make([]map[string]any, 0, len(nodes))
 	for _, n := range nodes {
-		kind, ver, flags, bytes := "missing", uint64(0), uint32(0), 0
+		kind, ver, flags, bytes := "missing", uint64(0), uint64(0), 0
 		role := "other"
 		if n.Engine != nil {
 			v := n.Engine.LocalView(ks, key)
